backend/requirements.txt
fastapi
uvicorn[standard]
xai-sdk
python-dotenv